refactor(//cpp/bin/torchtrtc): Refactor the CLI to make it easier to extend later
